pyproject.toml
[project]
name = "repl-nix-sight-undercover"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "opencv-python>=4.10.0.84",
    "numpy>=2.1.3",
    "pillow>=11.0.0",
    "werkzeug>=3.1.3",
    "flask-limiter>=3.9.2",
    "flask-cors>=5.0.0",
    "gunicorn>=23.0.0",
]
